Windows-App/requirements.txt
# UI
PyQt6>=6.6

# USB (shared with Windows-Script)
pyusb>=1.2
libusb-package>=1.0

# Transcription
openai-whisper>=20240930
torch>=2.1
torchaudio>=2.1

# Mic trigger (Windows audio session API)
pycaw>=20240210
comtypes>=1.4

# Build
pyinstaller>=6.0
